% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aln.r
\name{feats_to_aln}
\alias{feats_to_aln}
\title{Match external features to alignment}
\usage{
feats_to_aln(aln, feats)
}
\arguments{
\item{aln}{Alignment in tibble form. Output from seq_pos_from_aln.
Must have columns 'seq_id', 'seq_pos'.}

\item{feats}{Feature table. Must have columns 'seq_id', 'feat_id',
'start', 'end'. It should not have overlapping features in the same
sequence.}
}
\value{

}
\description{
For internal use.
}
