% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aln.r
\name{plot_aln_annots}
\alias{plot_aln_annots}
\title{Plot alignment with features}
\usage{
plot_aln_annots(aln, feats, region = NULL)
}
\arguments{
\item{aln}{Alignment in fasta format from seqinr. See
examples.}

\item{feats}{Feature table. Must have columns 'seq_id', 'feat_id',
'start', 'end'. It should not have overlapping features in the same}

\item{region}{Start and end positions of range to plot. If NULL, plot
all the alignment.}
}
\value{
A ggplot2 plot
}
\description{
Takes an alignment in fasta format as defined in
seqinr, and a tibble of features. Produces plot with
annotated features.
}
