% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{read_ipro_tsv}
\alias{read_ipro_tsv}
\title{Read InterPro 5 TSV output file}
\usage{
read_ipro_tsv(file, date_format = "\%d-\%m-\%Y")
}
\arguments{
\item{file}{TSV file from InterPro 5}

\item{date_format}{format for parsing date}
}
\value{
A tibble
}
\description{
Variable row lengths cause warnings that can be ignored.
}
