% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aln.r
\name{seq_pos_from_aln}
\alias{seq_pos_from_aln}
\title{Original sequence position from aligned sequence}
\usage{
seq_pos_from_aln(seq)
}
\arguments{
\item{seq}{An aligned sequence in vector form. Gaps as "-".}
}
\value{
A tibble
}
\description{
Internal use.
}
